requirements.txt
python-dotenv==0.20.*
BeautifulSoup4==4.11.1
lxml==4.9.1
evernote3==1.25.14
python-oauth2==1.1.1